Move policy loading out of generic_agent.h into loading.h
