+++ cf-agent/cf-agent.c
@@ -73,6 +73,7 @@
 #include <bootstrap.h>
 #include <misc_lib.h>
 #include <buffer.h>
+#include <loading.h>
 
 #include <mod_common.h>
 
@@ -239,7 +240,7 @@ int main(int argc, char *argv[])
     Policy *policy = NULL;
     if (GenericAgentCheckPolicy(config, ALWAYS_VALIDATE, true))
     {
-        policy = GenericAgentLoadPolicy(ctx, config);
+        policy = LoadPolicy(ctx, config);
     }
     else if (config->tty_interactive)
     {
@@ -250,7 +251,7 @@ int main(int argc, char *argv[])
         Log(LOG_LEVEL_ERR, "CFEngine was not able to get confirmation of promises from cf-promises, so going to failsafe");
         EvalContextClassPutHard(ctx, "failsafe_fallback", "attribute_name=Errors,source=agent");
         GenericAgentConfigSetInputFile(config, GetInputDir(), "failsafe.cf");
-        policy = GenericAgentLoadPolicy(ctx, config);
+        policy = LoadPolicy(ctx, config);
     }
     assert(policy);
