+++ cf-serverd/cf-serverd.c
@@ -23,9 +23,10 @@
 */
 
 #include <cf-serverd-functions.h>
-#include <known_dirs.h>
 
+#include <known_dirs.h>
 #include <server_transform.h>
+#include <loading.h>
 
 int main(int argc, char *argv[])
 {
@@ -38,7 +39,7 @@ int main(int argc, char *argv[])
     Policy *policy = NULL;
     if (GenericAgentCheckPolicy(config, false, false))
     {
-        policy = GenericAgentLoadPolicy(ctx, config);
+        policy = LoadPolicy(ctx, config);
     }
     else if (config->tty_interactive)
     {
@@ -49,7 +50,7 @@ int main(int argc, char *argv[])
         Log(LOG_LEVEL_ERR, "CFEngine was not able to get confirmation of promises from cf-promises, so going to failsafe");
         EvalContextClassPutHard(ctx, "failsafe_fallback", "attribute_name=Errors,source=agent");
         GenericAgentConfigSetInputFile(config, GetInputDir(), "failsafe.cf");
-        policy = GenericAgentLoadPolicy(ctx, config);
+        policy = LoadPolicy(ctx, config);
     }
 
     ThisAgentInit();
